refactor: :recycle: Separa os arquivos do redux em uma pasta
